add all missing image_method decorators
