+++ ants/core/__init__.py
@@ -13,9 +13,6 @@ from .ants_image import (copy_image_info,
                          get_direction,
                          set_spacing,
                          get_spacing,
-                         image_physical_space_consistency,
-                         image_type_cast,
-                         allclose,
                          is_image,
                          from_pointer)
 from .ants_metric_io import (new_ants_metric,
