+++ ants/ops/iMath.py
@@ -61,7 +61,7 @@ _iMathOps = {'FillHoles',
             'Sharpen',
             'PropagateLabelsThroughMask'}
 
-
+@image_method
 def multiply_images(image1, image2):
     return image1 * image2
 
@@ -111,67 +111,87 @@ image_math = iMath
 def iMath_ops():
     return _iMathOps
 
-
+@image_method
 def iMath_canny(image, sigma, lower, upper):
     return iMath(image, 'Canny', sigma, lower, upper)
 
+@image_method
 def iMath_distance_map(image, use_spacing=True):
     return iMath(image, 'DistanceMap', use_spacing)
 
+@image_method
 def iMath_fill_holes(image, hole_type=2):
     return iMath(image, 'FillHoles', hole_type)
 
+@image_method
 def iMath_GC(image, radius=1):
     return iMath(image, 'GC', radius)
 
+@image_method
 def iMath_GD(image, radius=1):
     return iMath(image, 'GD', radius)
 
+@image_method
 def iMath_GE(image, radius=1):
     return iMath(image, 'GE', radius)
 
+@image_method
 def iMath_GO(image, radius=1):
     return iMath(image, 'GO', radius)
 
+@image_method
 def iMath_get_largest_component(image, min_size=50):
     return iMath(image, 'GetLargestComponent', min_size)
 
+@image_method
 def iMath_grad(image, sigma=0.5, normalize=False):
     return iMath(image, 'Grad', sigma, normalize)
 
+@image_method
 def iMath_histogram_equalization(image, alpha, beta):
     return iMath(image, 'HistogramEqualization', alpha, beta)
 
+@image_method
 def iMath_laplacian(image, sigma=0.5, normalize=False):
     return iMath(image, 'Laplacian', sigma, normalize)
 
+@image_method
 def iMath_MC(image, radius=1, value=1, shape=1, parametric=False, lines=3, thickness=1, include_center=False):
     return iMath(image, 'MC', radius, value, shape, parametric, lines, thickness, include_center)
 
+@image_method
 def iMath_MD(image, radius=1, value=1, shape=1, parametric=False, lines=3, thickness=1, include_center=False):
     return iMath(image, 'MD', radius, value, shape, parametric, lines, thickness, include_center)
 
+@image_method
 def iMath_ME(image, radius=1, value=1, shape=1, parametric=False, lines=3, thickness=1, include_center=False):
     return iMath(image, 'ME', radius, value, shape, parametric, lines, thickness, include_center)
 
+@image_method
 def iMath_MO(image, radius=1, value=1, shape=1, parametric=False, lines=3, thickness=1, include_center=False):
     return iMath(image, 'MO', radius, value, shape, parametric, lines, thickness, include_center)
 
+@image_method
 def iMath_maurer_distance(image, foreground=1):
     return iMath(image, 'MaurerDistance', foreground)
 
+@image_method
 def iMath_normalize(image):
     return iMath(image, 'Normalize')
 
+@image_method
 def iMath_pad(image, padding):
     return iMath(image, 'Pad', padding)
 
+@image_method
 def iMath_perona_malik(image, conductance=0.25, n_iterations=1):
     return iMath(image, 'PeronaMalik', conductance, n_iterations)
 
+@image_method
 def iMath_sharpen(image):
     return iMath(image, 'Sharpen')
 
+@image_method
 def iMath_propagate_labels_through_mask(image, labels, stopping_value=100, propagation_method=0):
     """
     >>> import ants
@@ -181,6 +201,7 @@ def iMath_propagate_labels_through_mask(image, labels, stopping_value=100, propa
     """
     return iMath(image, 'PropagateLabelsThroughMask', labels, stopping_value, propagation_method)
 
+@image_method
 def iMath_truncate_intensity(image, lower_q, upper_q, n_bins=64):
     """
     >>> import ants
