+++ ants/plotting/plot_ortho.py
@@ -24,7 +24,9 @@ from mpl_toolkits.axes_grid1.inset_locator import inset_axes
 
 import numpy as np
 import ants
+from ants.decorators import image_method
 
+@image_method
 def plot_ortho(
     image,
     overlay=None,
