+++ ants/registration/average_transform.py
