+++ ants/utils/ndimage_to_list.py
@@ -5,7 +5,9 @@ __all__ = ['ndimage_to_list',
 import numpy as np
 
 import ants
+from ants.decorators import image_method
 
+@image_method
 def list_to_ndimage( image, image_list ):
     """
     Merge list of multiple scalar ANTsImage types of dimension into one
@@ -59,7 +61,7 @@ def list_to_ndimage( image, image_list ):
     return newImage
 
 
-
+@image_method
 def ndimage_to_list(image):
     """
     Split a n dimensional ANTsImage into a list
